+++ mypy_protobuf/main.py
@@ -87,11 +87,6 @@ PROTO_ENUM_RESERVED = {
 }
 
 
-def _build_typevar_name(service_name: str, method_name: str) -> str:
-    # Prefix with underscore to avoid public api error: https://stackoverflow.com/a/78871465
-    return f"_{service_name}{method_name}Type"
-
-
 def _mangle_global_identifier(name: str) -> str:
     """
     Module level identifiers are mangled and aliased so that they can be disambiguated
@@ -180,7 +175,7 @@ class PkgWriter(object):
         eg. self._import("typing", "Literal") -> "Literal"
         """
         if path == "typing_extensions":
-            stabilization = {"TypeAlias": (3, 10), "TypeVar": (3, 13)}
+            stabilization = {"TypeAlias": (3, 10), "TypeVar": (3, 13), "type_check_only": (3, 12)}
             assert name in stabilization
             if not self.typing_extensions_min or self.typing_extensions_min < stabilization[name]:
                 self.typing_extensions_min = stabilization[name]
@@ -816,45 +811,28 @@ class PkgWriter(object):
             wl("...")
         wl("")
 
-    def write_grpc_type_vars(self, service: d.ServiceDescriptorProto) -> None:
+    def write_grpc_stub_methods(self, service: d.ServiceDescriptorProto, scl_prefix: SourceCodeLocation, *, is_async: bool, both: bool = False, ignore_assignment_errors: bool = False) -> None:
         wl = self._write_line
         methods = [(i, m) for i, m in enumerate(service.method) if m.name not in PYTHON_RESERVED]
         if not methods:
             return
-        for _, method in methods:
-            wl("{} = {}(", _build_typevar_name(service.name, method.name), self._import("typing_extensions", "TypeVar"))
-            with self._indent():
-                wl("'{}',", _build_typevar_name(service.name, method.name))
-                wl("{}[", self._callable_type(method, is_async=False))
-                with self._indent():
-                    wl("{},", self._input_type(method))
-                    wl("{},", self._output_type(method))
-                wl("],")
-                wl("{}[", self._callable_type(method, is_async=True))
-                with self._indent():
-                    wl("{},", self._input_type(method))
-                    wl("{},", self._output_type(method))
-                wl("],")
-                wl("default={}[", self._callable_type(method, is_async=False))
-                with self._indent():
-                    wl("{},", self._input_type(method))
-                    wl("{},", self._output_type(method))
-                wl("],")
-            wl(")")
-            wl("")
 
-    def write_self_types(self, service: d.ServiceDescriptorProto, is_async: bool) -> None:
-        wl = self._write_line
-        methods = [(i, m) for i, m in enumerate(service.method) if m.name not in PYTHON_RESERVED]
-        if not methods:
-            return
-        for _, method in methods:
-            with self._indent():
-                wl("{}[", self._callable_type(method, is_async=is_async))
-                with self._indent():
-                    wl("{},", self._input_type(method))
-                    wl("{},", self._output_type(method))
-                wl("],")
+        def type_str(method: d.MethodDescriptorProto, is_async: bool) -> str:
+            return f"{self._callable_type(method, is_async=is_async)}[{self._input_type(method)}, {self._output_type(method)}]"
+
+        for i, method in methods:
+            scl = scl_prefix + [d.ServiceDescriptorProto.METHOD_FIELD_NUMBER, i]
+            if both:
+                wl(
+                    "{}: {}[{}, {}]",
+                    method.name,
+                    self._import("typing", "Union"),
+                    type_str(method, is_async=False),
+                    type_str(method, is_async=True),
+                )
+            else:
+                wl("{}: {}{}", method.name, type_str(method, is_async=is_async), "" if not ignore_assignment_errors else "  # type: ignore[assignment]")
+            self._write_comments(scl)
 
     def write_grpc_methods(self, service: d.ServiceDescriptorProto, scl_prefix: SourceCodeLocation) -> None:
         wl = self._write_line
@@ -885,19 +863,6 @@ class PkgWriter(object):
                         wl("...")
             wl("")
 
-    def write_grpc_stub_methods(self, service: d.ServiceDescriptorProto, scl_prefix: SourceCodeLocation, is_async: bool = False) -> None:
-        wl = self._write_line
-        methods = [(i, m) for i, m in enumerate(service.method) if m.name not in PYTHON_RESERVED]
-        if not methods:
-            wl("...")
-            wl("")
-        for i, method in methods:
-            scl = scl_prefix + [d.ServiceDescriptorProto.METHOD_FIELD_NUMBER, i]
-
-            wl("{}: {}", method.name, f"{_build_typevar_name(service.name, method.name)}")
-            self._write_comments(scl)
-            wl("")
-
     def write_grpc_services(
         self,
         services: Iterable[d.ServiceDescriptorProto],
@@ -906,14 +871,15 @@ class PkgWriter(object):
         wl = self._write_line
         wl("GRPC_GENERATED_VERSION: str")
         wl("GRPC_VERSION: str")
+        wl("")
         for i, service in enumerate(services):
             if service.name in PYTHON_RESERVED:
                 continue
 
             scl = scl_prefix + [i]
 
-            # Type vars
-            self.write_grpc_type_vars(service)
+            class_name = f"{service.name}Stub"
+            async_class_alias = f"{service.name}AsyncStub"
 
             # The stub client
             if service.options.deprecated:
@@ -921,43 +887,44 @@ class PkgWriter(object):
                     scl + [d.ServiceDescriptorProto.OPTIONS_FIELD_NUMBER] + [d.ServiceOptions.DEPRECATED_FIELD_NUMBER],
                     "This stub has been marked as deprecated using proto service options.",
                 )
-            class_name = f"{service.name}Stub"
             wl(
-                "class {}({}[{}]):",
+                "class {}:",
                 class_name,
-                self._import("typing", "Generic"),
-                ", ".join(f"{_build_typevar_name(service.name, method.name)}" for method in service.method),
             )
             with self._indent():
                 if self._write_comments(scl):
                     wl("")
-
                 # Write sync overload
                 wl("@{}", self._import("typing", "overload"))
-                wl("def __init__(self: {}[", class_name)
-                self.write_self_types(service, False)
                 wl(
-                    "], channel: {}) -> None: ...",
+                    "def __new__(cls, channel: {}) -> {}: ...",
                     self._import("grpc", "Channel"),
+                    class_name,
                 )
-                wl("")
 
                 # Write async overload
                 wl("@{}", self._import("typing", "overload"))
-                wl("def __init__(self: {}[", class_name)
-                self.write_self_types(service, True)
                 wl(
-                    "], channel: {}) -> None: ...",
+                    "def __new__(cls, channel: {}) -> {}: ...",
                     self._import("grpc.aio", "Channel"),
+                    async_class_alias,
                 )
+                self.write_grpc_stub_methods(service, scl, is_async=False)
                 wl("")
 
-                self.write_grpc_stub_methods(service, scl)
-
-            # Write AsyncStub alias
-            wl("{}AsyncStub: {} = {}[", service.name, self._import("typing_extensions", "TypeAlias"), class_name)
-            self.write_self_types(service, True)
-            wl("]")
+            # Write AsyncStub
+            if service.options.deprecated:
+                self._write_deprecation_warning(
+                    scl + [d.ServiceDescriptorProto.OPTIONS_FIELD_NUMBER] + [d.ServiceOptions.DEPRECATED_FIELD_NUMBER],
+                    "This stub has been marked as deprecated using proto service options.",
+                )
+            wl("@{}", self._import("typing", "type_check_only"))
+            wl("class {}({}):", async_class_alias, class_name)
+            with self._indent():
+                if self._write_comments(scl):
+                    wl("")
+                wl("def __init__(self, channel: {}) -> None: ...", self._import("grpc.aio", "Channel"))
+                self.write_grpc_stub_methods(service, scl, is_async=True, ignore_assignment_errors=True)
             wl("")
 
             # The service definition interface